Add a prototype of the meeting specific page.
You can now tap on a search result and it will open the meeting
specific page, which is a tab view. Gestures still work. You can click
back to go to the previous page.

The tab view is implemented in Storyboard, but is instantiated in code.
It is given an ID which is referenced when a new meeting page is
created.