Added ability to view a file
When you click on a file in the Files page, it will load the file into
a web view. This works for a lot of file types, like pdf, txt, and
images.

Also added two sample files for testing, a pdf of our design mockups,
and a txt file of code.