Added Person specific page
When you tap a person on the Person page, the people page for a
specific meeting, or the search results, it opens a new page for that
person.

As of right now, it only displays their profile picture. It will need
to be expanded with company, title, and a list of meetings they are
attending.